# docs/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "ExportPdfRequest": {
        "properties": {
          "fit_text_to_page": {
            "default": true,
            "title": "Fit Text To Page",
            "type": "boolean"
          },
          "font_family": {
            "default": "Times-Roman",
            "enum": [
              "Times-Roman",
              "Helvetica",
              "Courier"
            ],
            "title": "Font Family",
            "type": "string"
          },
          "font_size": {
            "default": 12.0,
            "title": "Font Size",
            "type": "number"
          },
          "join_historical_line_breaks": {
            "default": true,
            "title": "Join Historical Line Breaks",
            "type": "boolean"
          },
          "layout_mode": {
            "default": "source-lines",
            "enum": [
              "source-lines",
              "reading"
            ],
            "title": "Layout Mode",
            "type": "string"
          },
          "line_spacing": {
            "default": 1.35,
            "title": "Line Spacing",
            "type": "number"
          },
          "margin_in": {
            "default": 0.8,
            "title": "Margin In",
            "type": "number"
          },
          "normalize_double_oblique_hyphen": {
            "default": false,
            "title": "Normalize Double Oblique Hyphen",
            "type": "boolean"
          },
          "normalize_long_s": {
            "default": false,
            "title": "Normalize Long S",
            "type": "boolean"
          },
          "normalize_low_double_quote": {
            "default": false,
            "title": "Normalize Low Double Quote",
            "type": "boolean"
          },
          "page_size": {
            "default": "letter",
            "enum": [
              "letter",
              "a4"
            ],
            "title": "Page Size",
            "type": "string"
          },
          "spread_mode": {
            "default": "split-spread",
            "enum": [
              "single",
              "split-spread"
            ],
            "title": "Spread Mode",
            "type": "string"
          }
        },
        "title": "ExportPdfRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/export/projects/{project_id}/pdf": {
      "get": {
        "operationId": "export_project_pdf_get_export_projects__project_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "layout_mode",
            "required": false,
            "schema": {
              "default": "source-lines",
              "enum": [
                "source-lines",
                "reading"
              ],
              "title": "Layout Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "font_family",
            "required": false,
            "schema": {
              "default": "Times-Roman",
              "enum": [
                "Times-Roman",
                "Helvetica",
                "Courier"
              ],
              "title": "Font Family",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "font_size",
            "required": false,
            "schema": {
              "default": 12.0,
              "title": "Font Size",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "line_spacing",
            "required": false,
            "schema": {
              "default": 1.35,
              "title": "Line Spacing",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "margin_in",
            "required": false,
            "schema": {
              "default": 0.8,
              "title": "Margin In",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "spread_mode",
            "required": false,
            "schema": {
              "default": "split-spread",
              "enum": [
                "single",
                "split-spread"
              ],
              "title": "Spread Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "join_historical_line_breaks",
            "required": false,
            "schema": {
              "default": true,
              "title": "Join Historical Line Breaks",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "normalize_low_double_quote",
            "required": false,
            "schema": {
              "default": false,
              "title": "Normalize Low Double Quote",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "normalize_long_s",
            "required": false,
            "schema": {
              "default": false,
              "title": "Normalize Long S",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "normalize_double_oblique_hyphen",
            "required": false,
            "schema": {
              "default": false,
              "title": "Normalize Double Oblique Hyphen",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "fit_text_to_page",
            "required": false,
            "schema": {
              "default": true,
              "title": "Fit Text To Page",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": "letter",
              "enum": [
                "letter",
                "a4"
              ],
              "title": "Page Size",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Project Pdf Get",
        "tags": [
          "export"
        ]
      },
      "post": {
        "operationId": "export_project_pdf_post_export_projects__project_id__pdf_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPdfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Project Pdf Post",
        "tags": [
          "export"
        ]
      }
    },
    "/export/projects/{project_id}/training-data": {
      "get": {
        "operationId": "export_project_training_data_export_projects__project_id__training_data_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Project Training Data",
        "tags": [
          "export"
        ]
      }
    },
...
  }
}